requirements.txt
Python==3.8.1
pytest==5.3.5
selenium==3.141.0
PyHamcrest==2.0.2
